feat: added ssl communication

+++ api/Utils/supervisorCheck.js
@@ -9,15 +9,18 @@ const {roles} =require('./constants')
 const supervisorCheck = async (employeeId,db,addSuperRole)=>{
     const emp = await db.Employee.findByPk(employeeId)
     const superRole = await db.Role.findOne({where:{role_name:roles[1]}})
-    const userRole = await db.UserRole.findOne({where:{roleId:superRole.id,userId:emp.userId}})
-    switch(addSuperRole)
-    {
-        case true:         
-            if(!userRole) await db.UserRole.create({roleId:superRole.id,userId:emp.userId})
-            break;
-        case false:
+    const superUserRole = await db.UserRole.findOne({where:{roleId:superRole.id,userId:emp.userId}})
+    if(addSuperRole){
+        if(!superUserRole) await db.UserRole.create({roleId:superRole.id,userId:emp.userId})
+    }else{
+        const labs = await db.Lab.findAll({where:{employeeId:employeeId}})
+        const workshops = await db.WorkshopSupervisor.findAll({where:{EmployeeId:employeeId}})
+        if(labs.length===0 && workshops.length===0){
+            console.log(emp)
             await db.UserRole.destroy({where:{roleId:superRole.id,userId:emp.userId}})
-            break;
-    } 
+        }
+            
+    }    
+
 }
 module.exports = supervisorCheck
